code clean up completed and responsibility of component saparated
